+++ libs/nest/src/lib/rabbit/rabbit.service.ts
@@ -3,6 +3,7 @@ import { Global, Injectable, OnModuleInit } from '@nestjs/common';
 import { BaseWorker } from '../worker';
 import { ConfigService } from '@nestjs/config';
 import { RabbitMessagePayload } from './rabbit.types';
+import { service_definitions } from '../../validation/generated/modelSchema/service_definitionsSchema';
 
 export enum WorkerTypes {
   PLATFORM = 'platform',
@@ -13,20 +14,19 @@ export enum WorkerTypes {
 
 export type RabbitMessageOptions = {
   exchange: string;
+  timeout?: number;
 };
 
 @Global()
 @Injectable()
 export class ColdRabbitService extends BaseWorker implements OnModuleInit {
-  public registered = false;
-  definition: any;
-
   constructor(private readonly config: ConfigService, private readonly client: AmqpConnection) {
     super(ColdRabbitService.name);
   }
 
   override async onModuleInit(): Promise<void> {
     await this.initializeExitHandlers();
+    //
   }
 
   /**
@@ -36,24 +36,33 @@ export class ColdRabbitService extends BaseWorker implements OnModuleInit {
    * @throws {Error} - If an unknown WorkerType is specified in the definition.
    * @param pkg
    */
-  public async register_service(pkg: { name: string; label: string; service_type: WorkerTypes; definition: any }): Promise<void> {
+  public async register_service(pkg: { name: string; label: string; service_type: WorkerTypes; definition: service_definitions }): Promise<service_definitions> {
     try {
-      this.definition = pkg.definition;
-
-      await this.publish(
+      this.logger.info('Registering service with COLD_API', { ...pkg });
+      const response = await this.request(
         'cold.service_definitions',
         {
-          name: pkg.name,
-          label: pkg.label,
-          type: pkg.service_type,
-          definition: pkg.definition,
+          data: {
+            name: pkg.name,
+            label: pkg.label,
+            type: pkg.service_type,
+            definition: pkg.definition,
+          },
+          event: 'service_started',
+          isRPC: true,
+          from: pkg.name,
         },
-        'service_started',
+        {
+          exchange: 'amq.direct',
+          timeout: 30000,
+        },
       );
 
-      this.registered = true;
+      return response;
     } catch (err: any) {
       this.logger.error(err.message, { error: err });
+
+      throw err;
     }
   }
 
@@ -70,7 +79,7 @@ export class ColdRabbitService extends BaseWorker implements OnModuleInit {
   }
 
   /**
-   * Publishes a message to a specified exchange with a given routing key.
+   * Publishes an async message to a specified exchange with a given routing key.
    *
    * @param {string} routingKey
    * @param {any} data
@@ -88,7 +97,7 @@ export class ColdRabbitService extends BaseWorker implements OnModuleInit {
     },
   ): Promise<void> {
     try {
-      await this.client?.publish(options.exchange, routingKey, {
+      await this.client.publish(options.exchange, routingKey, {
         data,
         event: event,
         from: this.config.getOrThrow('DD_SERVICE'),
@@ -104,27 +113,19 @@ export class ColdRabbitService extends BaseWorker implements OnModuleInit {
   }
 
   /***
-   * Create RPC message to rabbit
+   * Publish an RPC message to rabbit
    * @param {string} routingKey
    * @param body
    * @param {RabbitMessageOptions} options
    * @returns {any}
    * @throws {Error}
    */
-  public async request(
-    routingKey: string,
-    body: RabbitMessagePayload,
-    options: RabbitMessageOptions = {
-      exchange: 'amq.direct',
-    },
-  ): Promise<any> {
+  public async request(routingKey: string, body: RabbitMessagePayload, options?: RabbitMessageOptions): Promise<any> {
     try {
-      if (this.client?.managedConnection) {
-        return;
-      }
       const response = await this.client?.request({
-        exchange: options.exchange,
+        exchange: options?.exchange || 'amq.direct',
         routingKey: routingKey,
+        timeout: options?.timeout || 5000,
         payload: body,
       });